fixes bug in ActionKV.insert_but_ignore_index
The method no longer assumes the storage file is already pointing
at the end of the file.

Previously, this bug was covered up small storage sizes by the
BufReader used in `get`. When the entire file is smaller than the
buffer, any read to the file will result in the file's seek
position pointing to the end of the file.

My debug print statement revealed the bug because calling
BufReader.seek(SeekFrom::Current(0)) was resetting the
file position to the same location it would have been
without the buffer.